separated auth server and endpoints
