+++ API/endpoints.py
@@ -4,24 +4,17 @@ The endpoint called `endpoints` will return all available endpoints.
 """
 
 from http import HTTPStatus
-from flask import Flask, request, jsonify, _request_ctx_stack
+from flask import Flask
 from flask_restx import Resource, Api, reqparse
 from flask_cors import cross_origin
 import werkzeug.exceptions as wz
-import json
-import os
-from six.moves.urllib.request import urlopen
-from functools import wraps
-from jose import jwt
 
 import db.data as db
+import server
 
 app = Flask(__name__)
 api = Api(app)
 
-AUTH0_DOMAIN = os.environ.get("AUTH0_DOMAIN")
-API_AUDIENCE = os.environ.get("AUTH0_API_AUDIENCE")
-ALGORITHMS = ["RS256"]
 
 spotParser = reqparse.RequestParser()
 spotParser.add_argument('spotName', type=str, location='form')
@@ -41,114 +34,6 @@ factorParser.add_argument('factorAvailability', type=int, location='form')
 factorParser.add_argument('factorNoiseLevel', type=int, location='form')
 factorParser.add_argument('factorTemperature', type=int, location='form')
 factorParser.add_argument('factorAmbiance', type=int, location='form')
-
-
-# Error handler
-class AuthError(Exception):
-    def __init__(self, error, status_code):
-        self.error = error
-        self.status_code = status_code
-
-
-@app.errorhandler(AuthError)
-def handle_auth_error(ex):
-    response = jsonify(ex.error)
-    response.status_code = ex.status_code
-    return response
-
-
-# Format error response and append status code
-def get_token_auth_header():
-    """Obtains the Access Token from the Authorization Header
-    """
-    auth = request.headers.get("Authorization", None)
-    if not auth:
-        raise AuthError({"code": "authorization_header_missing",
-                        "description":
-                            "Authorization header is expected"}, 401)
-
-    parts = auth.split()
-
-    if parts[0].lower() != "bearer":
-        raise AuthError({"code": "invalid_header",
-                        "description":
-                            "Authorization header must start with"
-                            " Bearer"}, 401)
-    elif len(parts) == 1:
-        raise AuthError({"code": "invalid_header",
-                        "description": "Token not found"}, 401)
-    elif len(parts) > 2:
-        raise AuthError({"code": "invalid_header",
-                        "description":
-                            "Authorization header must be"
-                            " Bearer token"}, 401)
-
-    token = parts[1]
-    return token
-
-
-def requires_auth(f):
-    """Determines if the Access Token is valid
-    """
-    @wraps(f)
-    def decorated(*args, **kwargs):
-        token = get_token_auth_header()
-        jsonurl = urlopen("https://"+AUTH0_DOMAIN+"/.well-known/jwks.json")
-        jwks = json.loads(jsonurl.read())
-        unverified_header = jwt.get_unverified_header(token)
-        rsa_key = {}
-        for key in jwks["keys"]:
-            if key["kid"] == unverified_header["kid"]:
-                rsa_key = {
-                    "kty": key["kty"],
-                    "kid": key["kid"],
-                    "use": key["use"],
-                    "n": key["n"],
-                    "e": key["e"]
-                }
-        if rsa_key:
-            try:
-                payload = jwt.decode(
-                    token,
-                    rsa_key,
-                    algorithms=ALGORITHMS,
-                    audience=API_AUDIENCE,
-                    issuer="https://"+AUTH0_DOMAIN+"/"
-                )
-            except jwt.ExpiredSignatureError:
-                raise AuthError({"code": "token_expired",
-                                "description": "token is expired"}, 401)
-            except jwt.JWTClaimsError:
-                raise AuthError({"code": "invalid_claims",
-                                "description":
-                                    "incorrect claims,"
-                                    "please check the audience and issuer"}, 401)  # noqa: E501
-            except Exception:
-                raise AuthError({"code": "invalid_header",
-                                "description":
-                                    "Unable to parse authentication"
-                                    " token."}, 401)
-
-            _request_ctx_stack.top.current_user = payload
-            return f(*args, **kwargs)
-        raise AuthError({"code": "invalid_header",
-                        "description": "Unable to find appropriate key"}, 401)
-    return decorated
-
-
-def requires_scope(required_scope):
-    """Determines if the required scope is present in the Access Token
-    Args:
-        required_scope (str): The scope required to access the resource
-    """
-    token = get_token_auth_header()
-    unverified_claims = jwt.get_unverified_claims(token)
-    if unverified_claims.get("scope"):
-        token_scopes = unverified_claims["scope"].split()
-        for token_scope in token_scopes:
-            if token_scope == required_scope:
-                return True
-    return False
 
 
 @api.route('/hello')
@@ -185,7 +70,7 @@ class Spot(Resource):
     @api.doc(parser=spotParser)
     @cross_origin(headers=["Content-Type", "Authorization"])
     @cross_origin(headers=["Access-Control-Allow-Origin", "http://localhost:3000"])  # noqa: E501
-    @requires_auth
+    @server.requires_auth
     def post(self):
         """
         Creates a new spot
@@ -225,7 +110,7 @@ class SpotDetail(Resource):
     @api.doc(parser=spotParser)
     @cross_origin(headers=["Content-Type", "Authorization"])
     @cross_origin(headers=["Access-Control-Allow-Origin", "http://localhost:3000"])  # noqa: E501
-    @requires_auth
+    @server.requires_auth
     def put(self, spot_id):
         """
         Update a spot
@@ -247,7 +132,7 @@ class SpotDetail(Resource):
     @api.response(HTTPStatus.NOT_FOUND, 'Not Found')
     @cross_origin(headers=["Content-Type", "Authorization"])
     @cross_origin(headers=["Access-Control-Allow-Origin", "http://localhost:3000"])  # noqa: E501
-    @requires_auth
+    @server.requires_auth
     def delete(self, spot_id):
         """
         Delete a spot
@@ -269,7 +154,7 @@ class SpotUpdateFactor(Resource):
     @api.doc(parser=factorParser)
     @cross_origin(headers=["Content-Type", "Authorization"])
     @cross_origin(headers=["Access-Control-Allow-Origin", "http://localhost:3000"])  # noqa: E501
-    @requires_auth
+    @server.requires_auth
     def put(self, spot_id):
         """
         Update a spot factor
@@ -289,7 +174,7 @@ class Review(Resource):
     @api.doc(parser=reviewParser)
     @cross_origin(headers=["Content-Type", "Authorization"])
     @cross_origin(headers=["Access-Control-Allow-Origin", "http://localhost:3000"])  # noqa: E501
-    @requires_auth
+    @server.requires_auth
     def post(self):
         """
         Creates a new review
@@ -311,7 +196,7 @@ class ReviewDetail(Resource):
     @api.response(HTTPStatus.NOT_ACCEPTABLE, 'A duplicate key')
     @cross_origin(headers=["Content-Type", "Authorization"])
     @cross_origin(headers=["Access-Control-Allow-Origin", "http://localhost:3000"])  # noqa: E501
-    @requires_auth
+    @server.requires_auth
     def delete(self, review_id):
         """
         Deletes a new review
